Remove API key from image response
The Weather Underground API will return an image URL that contains the
provided API key in the query string when requesting the `radar` or
`satellite` endpoints. This is not necessary information (and really
shouldn't be in there), so this PR removes it from the response and
replaces it with a benign `_render=image`.